requirements.txt
baidu_aip==4.16.6
docx==0.2.4
Pillow==9.2.0
python_docx==0.8.11
xlrd==1.2.0
xlwt==1.3.0
